Add some CLI options.
My main concern was allowing me to set it up for publishing to an organisation, without overwriting my default answer for username.

I've also added options for CLI support, coverage, and coveralls.

There are no tests associated with this, only because I don't know how to test CLI options (if someone points me in the right direction, I will add tests). It would be good to be able to test *if* a question got asked (not sure if that is possible), since there is now some logic in the `when` conditions.

Finally, the use of `self` is not needed with arrow functions, `this` is bound to the surrounding scope. So I deleted all `self` references in favor of `this`.

Closes #51